% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.test.R
\name{cnm.test}
\alias{cnm.test}
\title{Co-occurrence Null Model test.}
\usage{
cnm.test(com, method = "r1", nits = 5000, burn = 500, thin = 10,
  threshold = 0, null.out = FALSE)
}
\arguments{
\item{com}{Community matrix with species in columns.}

\item{method}{Permutation method. Options include: r00 = fully random, r0 =
maintain site frequencies, r1 = maintain both site and species frequencies.
For more options see the oecosimu function in the vegan package.}

\item{nits}{Number of iterations to conduct.}

\item{burn}{The number permutations to conduct prior to recording matrices.}

\item{thin}{The number of discarded matrices between iterations.}

\item{threshold}{Level of abundance necessary for inclusion in the modeling
process. This allows for the removal of some species that are at levels of
abundance that are not likely to be ecologically relevant.}

\item{null.out}{Logical: should the null modeled set of communities be
returned.}
}
\description{
Provides high level user access to the null modeling tools in the vegan
package.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
