% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_net.R
\name{cond_net}
\alias{cond_net}
\title{Conditional probability network.}
\usage{
cond_net(x)
}
\arguments{
\item{x}{A matrix of species co-occurrences with species in columns
and observations in rows.}
}
\value{
Returns a matrix of conditional probabilities for a set of species.
}
\description{
Generates a matrix of conditional probabilities for a set of
species co-occurrences.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
