% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_prob.R
\name{cond_prob}
\alias{cond_prob}
\title{Conditional probability calculator.}
\usage{
cond_prob(a, b)
}
\arguments{
\item{a}{Occurrences of a species.}

\item{b}{Occurrences of another species.}
}
\value{
Returns the conditional probabilities for each species
based on the other species
}
\description{
Calculates conditional probabilities using co-occurrences of two species.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
