% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coNet.R
\name{coNet}
\alias{coNet}
\title{Co-occurrence based interaction network modeling.}
\usage{
coNet(x = "co-occurrence matrix", ci.p = 95, raw = FALSE)
}
\arguments{
\item{x}{A co-occurrence matrix with observations in rows and
species in columns.}

\item{ci.p}{Interval used for edge removal in percent (e.g. use 95
for a nintey-five percent confidence interval).}

\item{raw}{LOGICAL: should the original matrix of conditional
probabilities, prior to removel of conditional probabilities
that are within the removal interval.}
}
\value{
An interaction network model in matrix form with
    "non-significant" links removed and relativized to the joint
    probabilities (DEFAULT) or not (raw = TRUE). If relativized,
    the matrix is the deviations of the conditional probabilities
    from the joint probabilities. For conditional probabilities
    equal to the joint probabilities, this value is 0. This
    value can also range from 1 to -1, depending on the magnitude
    of the difference between the conditional and joint
    probabilities.
}
\description{
Generates an interaction network model based on a matrix of
co-occurrences (i.e. repeated observations in space). This function
is based on the method described in
\insertCite{Araujo2011}{conetto} updated to use conditional
probabilities.
}
\note{
Given a set of repeated observations of variables
    (e.g. biological species), a network of model of
    interdependencies is estimated using conditional probabilties
    (\eqn{P(S_i|S_j)}). This is calculated using Bayes' Theorem, as
    \eqn{P(S_i|S_j) = \frac{P(S_i,S_j)}{P(S_j)}}. \eqn{P(S_i,S_j)}
    is the joint probability, the probability of observing
    species (\eqn{S_i} and \eqn{S_j}), which is calculated from the
    individual probabilities of each species (\eqn{P(S)}). The
    total abundance of each species is used to quantify the
    individual probabilities of each species, such the \eqn{P(S_i)
    = \frac{S_i}{N}}, where \eqn{N} is the total number of
    observational units. The joint probabilities are similarly
    calculated as the total number of co-occurrences divided by the
    total number of observational units, \eqn{P(S_i,S_j) =
    \frac{(S_i,S_j)}{N}}. For more details, such as the interval
    based test, see \insertCite{Araujo2011}{conetto}.
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
