% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPNT.R
\name{CPNT}
\alias{CPNT}
\title{CPNT}
\usage{
CPNT(data1, data2, num, formula)
}
\arguments{
\item{data1}{a data.frame in which to represent  the relationship among nodes:if one node links another, the relationship between them is 1, otherwise is 0}

\item{data2}{a data.frame in which includes all variables needed in formula}

\item{num}{is the number of nodes}

\item{formula}{a formula to be passed to function coxph(). For more details, please refer to package survival.}
}
\value{
a data.frame
}
\description{
CPNT
}
\examples{
library(survival)
data("data1")
data("data2")
num <- dim(data1[1])
formula <-Surv(ctime,censor)~V1+V2+V3+V4+V5+V6+V7+V1_2+V2_3+V3_4+V3_5+V3_6+V3_7+PC1+PC2+PC3+PC4+PC5
result <- CPNT(data1, data2, num, formula)
}
