% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanNet.R
\name{meanNet}
\alias{meanNet}
\title{Calculate the mean of a set of networks.}
\usage{
meanNet(x, zero.na = TRUE)
}
\arguments{
\item{x}{A list of networks to be used for calculating distances.}

\item{zero.na}{LOGICAL: Should NA values be treated as zeros?}
}
\value{
Returns a matrix of means.
}
\description{
Given a set of networks in matrix format, this function will
calculate the cell-wise (Hadamard) mean.
}
\examples{

net.l <- lapply(1:10, function(x) matrix(runif(100), nrow = 10))
meanNet(net.l)

}
\seealso{
\code{\link{coNet}} %% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
