% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minNet.R
\name{minNet}
\alias{minNet}
\title{Minimize a network using any network generating function and an abundance threshold}
\usage{
minNet(x, FUN = "cdNet", alpha = 0.05)
}
\arguments{
\item{x}{A network in matrix format.}

\item{FUN}{Function to use for minimization.}

\item{alpha}{Significance level.}
}
\value{
The minimized network is returned in matrix format.
}
\description{
Method for reducing the number of edges in a network.
}
\details{
Using a chosen minimization algorithm, a minimzed network is computed.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
