% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMean.R
\name{netMean}
\alias{netMean}
\title{Calculate the mean of a set of networks.}
\usage{
netMean(x, zero.na = TRUE)
}
\arguments{
\item{x}{A list of networks to be used for calculating distances.}

\item{zero.na}{LOGICAL: Should NA values be treated as zeros?}

\item{method}{Distance/dissimilarity method to use: "euclidean" or
"bray" for Bray-Curtis.}
}
\value{
Returns a distance objects of pairwise distance or
dissimilarities for all networks.
}
\description{
Given a set of networks in matrix format, this function will
calculate the cell-wise (Hadimard) mean.
}
\details{
This function computes the Euclidean (i.e. straight line) distance using all
edges for all pairs of networks in a list of networks. The mathematics of
using Euclidean distance to summarize the difference between two networks
has not been explored fully, so use with caution.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
