% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percThreshold.R
\name{percThreshold}
\alias{percThreshold}
\title{Calculates the percolation threshold.}
\usage{
percThreshold(x = "network matrix", step.size = 0.01)
}
\arguments{
\item{x}{A network in matrix form.}

\item{step.size}{The increment to use in determining the threshold.}
}
\description{
Determine the minimal network using the percolation threshold
method from \insertCite{Araujo2011}{conetto}.
}
\note{
%% ~~further notes~~
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
