% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_net.R
\name{cond_net}
\alias{cond_net}
\title{Conditional probability network.}
\usage{
cond_net(x)
}
\arguments{
\item{x}{A matrix of species co-occurrences with species in columns
and observations in rows.}
}
\value{
Returns a matrix of conditional probabilities for a set of
    species.
}
\description{
Generates a matrix of conditional probabilities for a set of
co-occurrences.
}
\note{
This is primarily a low-level function used by coNet, it may
    be of interest to more advanced users who would like to
    directly compute the conditional probabilities from a set of
    co-occurrences.
}
\examples{

A <- c(1, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1)
B <- c(1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1)
C <- c(1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1)
D <- c(0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0)
M <- data.frame(A, B, C, D)
cond_net(M)

}
\seealso{
\code{\link{coNet}}, \code{\link{cond_prob}} %% ~~objects
    to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
