% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_prob.R
\name{cond_prob}
\alias{cond_prob}
\title{Conditional probability calculator.}
\usage{
cond_prob(a, b)
}
\arguments{
\item{a}{A set of repeated observations of occurrences.}

\item{b}{Another set of repeated observations of occurrences.}
}
\value{
Returns the conditional probabilities.
}
\description{
Calculates conditional probabilities using co-occurrences using
Bayes' Theorem, such that \eqn{P(A|B) = \frac{P(A,B)}{P(B)}}, where
\eqn{P(A,B) = \frac{(A,B)}{N}}, \eqn{P(B) = \frac{(B)}{N}} and
\eqn{N} is the total number of observational units.
}
\note{
This is primarily a low-level function used by coNet, it may
    be of interest to more advanced users.
}
\examples{

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
