% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coNet.R
\name{coNet}
\alias{coNet}
\title{Co-occurrence based interaction network modeling.}
\usage{
coNet(x = "co-occurrence matrix", ci.p = 95, raw = FALSE)
}
\arguments{
\item{x}{A co-occurrence matrix with observations in rows and
species in columns.}

\item{ci.p}{Interval used for edge removal in percent (e.g. use 95
for a nintey-five percent confidence interval).}

\item{raw}{LOGICAL: should the original matrix of conditional
probabilities, prior to removel of conditional probabilities
that are within the removal interval.}
}
\value{
An interaction network model in matrix form.
}
\description{
Generates an interaction network model based on a matrix of
co-occurrences (i.e. repeated observations in space). This function
is based on the method described in
\insertCite{Araujo2011}{conetto} updated to use conditional
probabilities.
}
\note{
%% ~~further notes~~
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
