% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percThreshold.R
\name{percThreshold}
\alias{percThreshold}
\title{Calculates the percolation threshold.}
\usage{
percThreshold(x = "network matrix", step.size = 0.01)
}
\arguments{
\item{x}{A network in matrix form.}

\item{step.size}{The increment to use in determining the threshold.}
}
\description{
Calculates the percolation threshold to determine the minimal network using
the methods of Araujo et al. 2011.
}
\details{
This function is a dependency for the co.net and CoNetwork functions.
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x = "network matrix", step.size = 0.01) 
{
    no.c <- no.clusters(graph.adjacency(x, weighted = TRUE))
    step <- 1
    while (no.c == 1) {
        x[x <= (step * step.size)] <- 0
        no.c <- no.clusters(graph.adjacency(x, weighted = TRUE))
        step <- step + 1
    }
    out <- list(threshold = ((step - 1) * step.size), isolated.nodes = colnames(x)[clusters(graph.adjacency(x, 
        weighted = TRUE))$membership == 2])
    return(out)
  }

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Matthew K. Lau
}
\keyword{~kwd1}
\keyword{~kwd2}
